Refactor DataFlowValueFactory implementation into several separate files
